android/requirements.txt
# androidhelper
paho-mqtt
python-dotenv
